First compiling version of audi integration in dcgp
