+++ src/CMakeLists.txt
@@ -3,7 +3,6 @@ SET(dCGP_LIB_SRC_LIST
 	${CMAKE_CURRENT_SOURCE_DIR}/expression.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/function_set.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/rng.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/wrapped_functions.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/fitness_functions.cpp
 )
