+++ src/dcgp.h
@@ -2,8 +2,6 @@
 #define DCGP_H
 
 #include "expression.h"
-#include "basis_function.h"
-#include "wrapped_functions.h"
 #include "fitness_functions.h"
 #include "function_set.h"
 #include "std_overloads.h"
