+++ src/function_set.cpp
@@ -1,4 +1,4 @@
-#include <stdexcept>
+#include <audi/gdual.hpp>
 
 #include "function_set.h"
 #include "wrapped_functions.h"
@@ -18,17 +18,13 @@ function_set::function_set(const std::vector<std::string>& list)
 void function_set::push_back(const std::string& function_name)
 {
     if (function_name=="sum")
-        m_functions.emplace_back(my_sum,d_my_sum,print_my_sum, function_name);
+        m_functions.emplace_back(my_sum<double>, my_sum<audi::gdual>, print_my_sum, function_name);
     else if (function_name=="diff")
-        m_functions.emplace_back(my_diff,d_my_diff,print_my_diff, function_name);
+        m_functions.emplace_back(my_diff<double>,my_diff<audi::gdual>,print_my_diff, function_name);
     else if (function_name=="mul")
-        m_functions.emplace_back(my_mul,d_my_mul,print_my_mul, function_name);
+        m_functions.emplace_back(my_mul<double>,my_mul<audi::gdual>,print_my_mul, function_name);
     else if (function_name=="div")
-        m_functions.emplace_back(my_div,d_my_div,print_my_div, function_name);
-    else if (function_name=="sqrt")
-        m_functions.emplace_back(my_sqrt,d_my_sqrt,print_my_sqrt, function_name);
-    else if (function_name=="pow")
-        m_functions.emplace_back(my_pow,d_my_pow,print_my_pow, function_name);
+        m_functions.emplace_back(my_div<double>,my_div<audi::gdual>,print_my_div, function_name);
     else 
         throw std::invalid_argument("Unimplemented function " + function_name);
 }
